+++ src/modelling/multibody/centroidal-momentum.cpp
@@ -1,0 +1,8 @@
+#include "aligator/modelling/multibody/centroidal-momentum.hpp"
+
+namespace aligator {
+
+template struct CentroidalMomentumResidualTpl<context::Scalar>;
+template struct CentroidalMomentumDataTpl<context::Scalar>;
+
+} // namespace aligator