+++ src/modelling/multibody/multibody-wrench-cone.cpp
@@ -1,0 +1,10 @@
+#include "aligator/modelling/multibody/multibody-wrench-cone.hpp"
+#ifdef ALIGATOR_PINOCCHIO_V3
+namespace aligator {
+
+template struct MultibodyWrenchConeResidualTpl<context::Scalar>;
+template struct MultibodyWrenchConeDataTpl<context::Scalar>;
+
+} // namespace aligator
+
+#endif